pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess_reasoning"
version = "0.1.0"
description = "Research pipeline for human vs LLM chess reasoning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "python-chess>=1.999",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "zstandard>=0.22"
]

[project.scripts]
chess-reasoning = "chess_reasoning.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
